requirements.txt
asgiref==3.7.2
attrs==23.1.0
bcrypt==4.0.1
black==23.10.0
blessed==1.20.0
botocore==1.31.57
cement==2.8.2
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.3
coverage==6.5.0
coveralls==3.3.1
cryptography==41.0.4
distlib==0.3.7
distro==1.8.0
Django==4.2.6
django-extensions==3.2.3
djangorestframework==3.14.0
docopt==0.6.2
filelock==3.12.4
flake8==6.1.0
googlemaps==4.10.0
haversine==2.8.0
identify==2.5.30
idna==3.4
jmespath==1.0.1
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
pathspec==0.10.1
Pillow==10.1.0
platformdirs==3.10.0
pre-commit==3.5.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
semantic-version==2.8.5
six==1.16.0
sqlparse==0.4.4
termcolor==1.1.0
texttable==1.6.7
tqdm==4.66.1
urllib3==1.26.16
virtualenv==20.24.5
wcwidth==0.1.9
websocket-client==0.59.0
